MBS-4517 Result type extracted to new :common:result module
