+++ subprojects/common/result/build.gradle.kts
@@ -1,0 +1,9 @@
+plugins {
+    id("convention.kotlin-jvm")
+    id("convention.publish-kotlin-library")
+    id("convention.libraries")
+}
+
+dependencies {
+    implementation(libs.funktionaleTry)
+}
