+++ subprojects/gradle/runner/service/build.gradle.kts
@@ -13,6 +13,7 @@ dependencies {
     api(project(":common:statsd"))
     api(project(":common:time"))
     implementation(project(":gradle:runner:shared"))
+    implementation(project(":common:result"))
     implementation(libs.funktionaleTry)
     implementation(libs.ddmlib)
     implementation(libs.rxJava)
