+++ subprojects/settings.gradle.kts
@@ -72,6 +72,7 @@ include(":common:okhttp")
 include(":common:file-storage")
 include(":common:test-okhttp")
 include(":common:report-viewer")
+include(":common:result")
 include(":common:elastic-logger")
 include(":common:elastic")
 include(":common:sentry-logger")
